Refactor bucket-related stuff into separate file
