+++ edsl/jobs/Jobs.py
@@ -15,7 +15,9 @@ from edsl.jobs.base import JobsRunnersRegistry, JobsRunnerDescriptor
 from edsl.jobs.Interview import Interview
 from edsl.coop.old import JobRunnerAPI, ResultsAPI
 
-from edsl.jobs.ModelBuckets import ModelBuckets, BucketCollection
+#from edsl.jobs.ModelBuckets import ModelBuckets, BucketCollection
+
+from edsl.jobs.buckets import BucketCollection
 
 class Jobs(Base):
     """
