+++ edsl/jobs/buckets.py
@@ -1,11 +1,7 @@
 import asyncio
 import time
-
-## TPM: Tokens-per-minute
-## RPM: Requests-per-minute 
-
-## TPM: 2,000,000
-## RPM:    10,000
+from collections import UserDict
+from matplotlib import pyplot as plt
 
 class TokenBucket:
     
@@ -63,7 +59,6 @@ class TokenBucket:
     
 
     def visualize(self):
-        from matplotlib import pyplot as plt
         times, tokens = zip(*self.get_log())
         start_time = times[0]
         times = [t - start_time for t in times]  # Normalize time to start from 0
@@ -75,22 +70,47 @@ class TokenBucket:
         plt.title('Token Bucket Usage Over Time')
         plt.show()
 
-# async def my_task(task_id, token_amount, bucket):
-#     await bucket.get_tokens(token_amount)  # Request specified number of tokens from the bucket
-#     print(f"Executing task {task_id} with {token_amount} tokens")
-#     # Simulate task execution
-#     await asyncio.sleep(1)
 
-# async def main():
-#     bucket = TokenBucket(capacity=10, refill_rate=2)  # Customize parameters as needed
+class ModelBuckets:
+    def __init__(self, requests_bucket: TokenBucket, tokens_bucket: TokenBucket):
+        self.requests_bucket = requests_bucket
+        self.tokens_bucket = tokens_bucket
 
-#     # Example tasks with varying token requirements
-#     tasks = [
-#         my_task(1, 3, bucket),
-#         my_task(2, 2, bucket),
-#         my_task(3, 5, bucket),
-#         my_task(4, 1, bucket)
-#     ]
-#     await asyncio.gather(*tasks)
+    def __add__(self, other):
+        return ModelBuckets(requests_bucket = self.requests_bucket + other.requests_bucket,
+                            tokens_bucket = self.tokens_bucket + other.tokens_bucket
+        )  
+    
+    def visualize(self):
+        plot1 = self.requests_bucket.visualize()
+        plot2 = self.tokens_bucket.visualize()
+        return plot1, plot2
 
-# asyncio.run(main())
+class BucketCollection(UserDict):
+    """When passed a model, will look up the associated buckets.
+    The keys are models, the value is a ModelBuckets 
+    """
+    def __init__(self):
+        super().__init__()
+
+    def add_model(self, model):
+        # compute the TPS and RPS from the model
+        TPS = model.TPM() / 60.0
+        RPS = model.RPM() / 60.0    
+        # create the buckets
+        requests_bucket = TokenBucket(capacity=2 * RPS, refill_rate=RPS)
+        tokens_bucket = TokenBucket(capacity=2 * TPS, refill_rate=TPS)
+        model_buckets = ModelBuckets(requests_bucket, tokens_bucket)
+        if model in self:
+            # it if already exists, combine the buckets
+            self[model] += model_buckets
+        else:          
+            self[model] = model_buckets
+
+    def visualize(self):
+        plots = {}
+        for model in self:
+            plots[model] = self[model].visualize()
+        return plots
+
+
